Refactor epic stories
